Moving files, better version management
As we want identify the installed version, we need need to specify the
appVersion in flathub in the build command. Another one is used for
Github Releases with fyne-cross.

But, at this time, we need a modified version of fyne-cross to ensure
that the "release" mode is set. It should be OK in a while.

As the release mode fails to work on seperated "main" package, we (for
now) moved the cmd/scoreboard content to the root.